module authentication-service
add GuestHelper, LogoutService, RelogService classes\nmove all service classes\nrename AuthService -> LoginService -> modified class\nmodify TokenHelper class\nmodify build.gradle
